refactor for clarity & best practices
